pyproject.toml
[tool.poetry]
name = "api-demo"
version = "0.1.0"
description = ""
authors = ["test <test@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
isort = "^5.13.2"
black = "^24.4.2"
mongoengine = "0.27.0"
pyyaml = "^6.0.1"
flask-cors = "4.0.0"
python-dateutil = "^2.9.0.post0"
pymongo = "4.0"
graphene = "2.1.9"
dnspython = "2.2.1"
flask-graphql = "2.0.1"
graphene-mongo = "0.1.8"
pytest = "^8.2.2"
requests = "^2.32.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
